+++ src/impls/$split-groups/__tests__/$split-groups.spec.ts
@@ -1,8 +1,18 @@
+import { $Promise } from '../../../../generate/async.macro.cjs';
+
 import assert from 'static-type-assert';
-import { $Iterable, $ResultIterable } from '../../../types/$iterable';
 import { $splitGroups } from 'iter-tools-es';
+import { $Iterable, $ResultIterable } from '../../../types/$iterable';
 
 declare const Ø: never;
+
+assert<$ResultIterable<[string, $ResultIterable<number>]>>(
+  $splitGroups(Ø as (x: number) => $Promise<string>)(Ø as $Iterable<number>),
+);
+
+assert<$ResultIterable<[string, $ResultIterable<number>]>>(
+  $splitGroups(Ø as (x: number) => $Promise<string>, Ø as $Iterable<number>),
+);
 
 assert<$ResultIterable<[string, $ResultIterable<string>]>>($splitGroups(Ø as string));